+++ lib/asset/sql.go
@@ -134,29 +134,6 @@ func (e *ErrorType) Scan(value interface{}) error {
 }
 
 // Value implements the driver.Valuer interface.
-// Simply returns the string representation of the ModelCategory.
-func (c *ModelCategory) Value() (driver.Value, error) {
-	return c.String(), nil
-}
-
-// Scan implements the sql.Scanner interface.
-// Simply decodes a string into the ModelCategory.
-func (c *ModelCategory) Scan(value interface{}) error {
-	s, ok := value.(string)
-	if !ok {
-		return errors.NewInternal("cannot scan model category: invalid string")
-	}
-
-	v, ok := ModelCategory_value[s]
-	if !ok {
-		return errors.NewInternal("cannot scan model category: unknown value")
-	}
-	*c = ModelCategory(v)
-
-	return nil
-}
-
-// Value implements the driver.Valuer interface.
 // Simply returns the string representation of the AssetKind.
 func (k *AssetKind) Value() (driver.Value, error) {
 	return k.String(), nil